Caldera e2e test (#63)
* Do not transfer all supply to L2, leave some for L1-L3 bridging test

* Bump blockscout pin to include ABI change

* Bump nitro contracts

* Disable rebuild of dev docker imgs w --build false

* Make UpgradeExecutor L3 arbowner

* Clarify logic around dev builds and rebuilding

The concepts of using the dev builds for nitro and blockscout and
rebuilding are now split into different variables internally to the
test-node.bash script and can be controlled independently.

The --dev flag by itself means to use the dev builds and to rebuild
them. --no-build-dev-nitro and -no-build-dev-blockscout disable
rebuilding the respective dev images. --no-build disables rebuilding the
dev images and also rebuilding contracts related images, and the node
images. This finer grained control saves time in the development cycle
by allowing the user to rebuild only what is needed.

An example command line to reinitialize the blockchain, use the dev
images, but not rebuild anything except blockscout would be:
./test-node.bash --init --dev --no-build --build-dev-blockscout --blockscout

Flags are read from left to right; flags to the right cancel
the effects of flags to the left.

* No skip_build

* Add comment

* Contracts and node images should rebuild on --init

* Declare variable before use for -u flag

* Change contracts to utils, add --force-build-utils

* Upgrade to the consensus-v31 release

* Update versions of contracts token-bridge and node

These updated versions are needed to work with the consensus-v31
release.

* fix: prysm config

* fix: remove irrelevant changes

* fix: pos

* fix: working config

* chore: reduce slot time to speed up deployment

* fix: enable electra since genesis

* fix: use newer geth

* fix: remove --config-name=interop

* fix: disable electra

* fix: use latest geth

* fix: add fee recipient

* fix: enable deneb

* fix: wait for tx sync

* fix: format

* feat: shorter slot for faster testing

* fix: readd dev flags

* fix: readd condition

* fix: reduce diff

* fix: docker compose

* ci: test pos

* ci: fix

* ci: debug log

* fix: also use 1337 for pos setup

* Test commit

* Test passing all the args through

* Test tokenbridge and l3node startups

* Changed the runner names

* Cleanup in ci bash

* Dont test for l3 tx if no l3node

* Added l3 node comment

* Add --l2-anytrust option to run in AnyTrust mode

test-node.bash can now be run with --l2-anytrust which runs the l2
nodes in AnyTrust mode. It creates a committee of 2 daserver with
assumed-honest set to 1, requiring successful stores of the batch data
to both of them. Nitro nodes sync from a mirror daserver. All dasevers
use local file storage for the batch data on their own volumes.

BLS keys for the committee are automatically generated and the keyset is
activated automatically on the SequencerInbox contract.

* Use latest Nitro release v3.1.2

* Added some testing for l3node with custom fee token

* Rename token6

* Set lower transfer amounts

* Update to nitro v3.2.1

* Update blockscout pin to arbitrum branch

* Add clabot config

* Fix PR comments

* merge upstream master

* fix test-node.bash

* add caldera smoke test

* add smoke test to ci

* fix smoke tests

* delete clabot

* remove unused vars

---------

Co-authored-by: Goran Vladika <goran.vladika@gmail.com>
Co-authored-by: Lee Bousfield <ljbousfield@gmail.com>
Co-authored-by: Tsahi Zidenberg <65945052+tsahee@users.noreply.github.com>
Co-authored-by: Tristan Wilson <twilson@offchainlabs.com>
Co-authored-by: Pepper Lebeck-Jobe <pepper@offchainlabs.com>
Co-authored-by: Pepper Lebeck-Jobe <eljobe@gmail.com>
Co-authored-by: gzeon <hng@offchainlabs.com>
Co-authored-by: gzeon <im@gzeon.dev>
Co-authored-by: Chris Buckland <cpbuckland88@gmail.com>
Co-authored-by: Tristan-Wilson <87238672+Tristan-Wilson@users.noreply.github.com>
Co-authored-by: Tsahi Zidenberg <tsahi@offchainlabs.com>
Co-authored-by: Harry Kalodner <harry@offchainlabs.com>